refactor(calendar): retire legacy calendar scrapers
