+++ src/ingestion/calendar/__init__.py
@@ -1,8 +1,1 @@
-"""Calendar data class — economic event calendars.
-
-Canonical imports::
-
-    from ingestion.calendar.scrapers.forexfactory import ForexFactoryCalendarClient
-    from ingestion.calendar.scrapers.investing import InvestingCalendarClient
-    from ingestion.calendar.scrapers.tradingeconomics import TradingEconomicsCalendarClient
-"""
+"""Calendar data class — economic and corporate event calendars."""
