--- src/ingestion/calendar/scrapers/__init__.py
@@ -1,1 +1,0 @@
-"""Calendar scrapers — ForexFactory, Investing.com, TradingEconomics."""
