--- src/ingestion/scrapers/forexfactory.py
@@ -1,1 +1,0 @@
-from ingestion.calendar.scrapers.forexfactory import *  # noqa: F401,F403
