--- src/ingestion/scrapers/investing.py
@@ -1,1 +1,0 @@
-from ingestion.calendar.scrapers.investing import *  # noqa: F401,F403
