--- src/ingestion/scrapers/tradingeconomics.py
@@ -1,1 +1,0 @@
-from ingestion.calendar.scrapers.tradingeconomics import *  # noqa: F401,F403
