+++ src/macro_data/service/_market.py
@@ -2,11 +2,7 @@ from __future__ import annotations
 
 from typing import Any
 
-from .base import (
-    LocalMacroDataServiceBase,
-    _VALID_MARKET_ASSET_CLASSES,
-    logger,
-)
+from .base import LocalMacroDataServiceBase
 
 
 class MarketOpsMixin(LocalMacroDataServiceBase):
@@ -53,25 +49,8 @@ class MarketOpsMixin(LocalMacroDataServiceBase):
         return {"total": len(rows), "prices": rows}
 
     def _op_fetch_live_markets(self, arguments: dict[str, Any]) -> dict[str, Any]:
-        from ingestion.scrapers import TradingEconomicsMarketsClient
-
-        asset_class = (arguments.get("asset_class") or "all").lower().strip()
-        try:
-            quotes = TradingEconomicsMarketsClient().fetch_markets()
-        except Exception as exc:
-            logger.warning("Live markets fetch failed: %s", exc)
-            return {"error": str(exc), "quotes": []}
-        items = [
-            {
-                "name": quote.name,
-                "asset_class": quote.asset_class,
-                "price": quote.price,
-                "change": quote.change,
-                "change_pct": quote.change_pct,
-                "symbol": quote.symbol,
-            }
-            for quote in quotes
-        ]
-        if asset_class != "all" and asset_class in _VALID_MARKET_ASSET_CLASSES:
-            items = [quote for quote in items if str(quote["asset_class"]).lower() == asset_class]
-        return {"total": len(items), "asset_class_filter": asset_class, "quotes": items}
+        del arguments
+        return {
+            "error": "live TradingEconomics market scraper retired in issue #123",
+            "quotes": [],
+        }
