+++ src/macro_data/service/base.py
@@ -52,9 +52,6 @@ def _document_family(doc: Any) -> str:
 _DOCUMENT_FAMILIES: frozenset[str] = frozenset(
     {"news", "note", "calendar", "release_report"}
 )
-_VALID_MARKET_ASSET_CLASSES = {
-    "index", "commodity", "fx", "bond", "stock", "crypto", "rate",
-}
 _VALID_RATE_TYPES = {"sofr", "effr", "obfr", "all"}
 _ARTICLE_DOMAIN_MAP: dict[str, str] = {
     "bloomberg.com": "bloomberg",
